move out helper functions
