+++ main.go
@@ -80,108 +80,6 @@ var defaultTimeZone = time.Local
 
 var logger mlog.Level
 
-// dateParamToEpoch turns a passed string parameter into an epoch we can send to the Zipper
-func dateParamToEpoch(s string, d int64) int32 {
-
-	if s == "" {
-		// return the default if nothing was passed
-		return int32(d)
-	}
-
-	// relative timestamp
-	if s[0] == '-' {
-
-		offset, err := intervalString(s, -1)
-		if err != nil {
-			return int32(d)
-		}
-
-		return int32(timeNow().Add(time.Duration(offset) * time.Second).Unix())
-	}
-
-	if s == "now" {
-		return int32(timeNow().Unix())
-	}
-
-	sint, err := strconv.Atoi(s)
-	if err == nil && len(s) > 8 {
-		return int32(sint) // We got a timestamp so returning it
-	}
-
-	if strings.Contains(s, "_") {
-		s = strings.Replace(s, "_", " ", 1) // Go can't parse _ in date strings
-	}
-
-	for _, format := range timeFormats {
-		t, err := time.ParseInLocation(format, s, defaultTimeZone)
-		if err == nil {
-			return int32(t.Unix())
-		}
-	}
-	return int32(d)
-}
-
-var errUnknownTimeUnits = errors.New("unknown time units")
-
-func intervalString(s string, defaultSign int) (int32, error) {
-
-	sign := defaultSign
-
-	switch s[0] {
-	case '-':
-		sign = -1
-		s = s[1:]
-	case '+':
-		sign = 1
-		s = s[1:]
-	}
-
-	var totalInterval int32
-	for len(s) > 0 {
-		var j int
-		for j < len(s) && '0' <= s[j] && s[j] <= '9' {
-			j++
-		}
-		var offsetStr string
-		offsetStr, s = s[:j], s[j:]
-
-		j = 0
-		for j < len(s) && (s[j] < '0' || '9' < s[j]) {
-			j++
-		}
-		var unitStr string
-		unitStr, s = s[:j], s[j:]
-
-		var units int
-		switch unitStr {
-		case "s", "sec", "secs", "second", "seconds":
-			units = 1
-		case "min", "mins", "minute", "minutes":
-			units = 60
-		case "h", "hour", "hours":
-			units = 60 * 60
-		case "d", "day", "days":
-			units = 24 * 60 * 60
-		case "w", "week", "weeks":
-			units = 7 * 24 * 60 * 60
-		case "mon", "month", "months":
-			units = 30 * 24 * 60 * 60
-		case "y", "year", "years":
-			units = 365 * 24 * 60 * 60
-		default:
-			return 0, errUnknownTimeUnits
-		}
-
-		offset, err := strconv.Atoi(offsetStr)
-		if err != nil {
-			return 0, err
-		}
-		totalInterval += int32(sign * offset * units)
-	}
-
-	return totalInterval, nil
-}
-
 func (z zipper) Find(metric string) (pb.GlobResponse, error) {
 
 	u, _ := url.Parse(string(z.z) + "/metrics/find/")
